requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data parsing
feedparser>=6.0.10  # For RSS feeds (ForkMonitor)

# Development/Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI
